requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.12.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
pydantic==1.10.0
cryptography==3.4.8
dnspython==2.2.0
